DL: refactor module imports
