+++ scirex/core/dl/__init__.py
@@ -1,4 +1,5 @@
 from .base import Network, Model
+from .nn import *
 from .mlp import MLP
 from .fcnn import FCNN
 from .tf_backend import *
