+++ scirex/core/dl/nn/__init__.py
@@ -1,5 +1,5 @@
-from .activation import *
-from .loss import *
-from .metrics import *
-from .others import *
-from .layers import *
+import .activations as activations
+import .losses as losses
+import .metrics as metrics
+import .layers as layers
+import .utils as utils
