+++ tests/scirex/core/dl/test_base.py
@@ -40,7 +40,7 @@ import jax.numpy as jnp
 import optax
 import equinox as eqx
 from scirex.core.dl import Model, Network
-from scirex.core.dl.nn import mse_loss
+from scirex.core.dl.losses import mse_loss
 
 
 key = jax.random.PRNGKey(0)
